+++ ImageItem/ImageItem.js
@@ -119,8 +119,8 @@ const ImageItemBase = kind({
 		})
 	},
 
-	render: ({captionPosition, children, css, disabled, src, ...rest}) => {
-		const [Component, marqueeProps] = (children && (captionPosition === 'below')) ? [MarqueeImageItem, {
+	render: ({captionPosition, children, css, disabled, orientation, src, ...rest}) => {
+		const [Component, marqueeProps] = (children && (orientation === 'horizontal' || captionPosition === 'below')) ? [MarqueeImageItem, {
 			alignment: 'center'
 		}] : [UiImageItem, null];
 
@@ -132,6 +132,7 @@ const ImageItemBase = kind({
 				css={css}
 				disabled={disabled}
 				imageComponent={ImageBase}
+				orientation={orientation}
 				src={src}
 			>
 				{children}